Relocate enchanting menus
