+++ src/main/java/dev/dpvb/survival/npc/enchanting/menus/EnchantingSelectionMenu.java
@@ -1,0 +1,4 @@
+package dev.dpvb.survival.npc.enchanting.menus;
+
+public class EnchantingSelectionMenu {
+}
